First fix vm

+++ VM/sources/apply_and.c
@@ -24,14 +24,14 @@ void				apply_and(t_process *process, t_env *env)
 	t_args		args[3];
 	uint32_t	reg;
 	uint32_t	pc;
-	uint32_t	val1;
-	uint32_t	val2;
+	int16_t		val1;
+	int16_t		val2;
 
 	decode(args, env->memory[process->pc], env->memory[process->pc + 1]);
 	pc = process->pc + BYPASS_ARG_ENCODE;
-	val1 = get_args(env->memory, pc, args[0].length);
+	val1 = (int16_t)get_args(env->memory, pc, args[0].length);
 	pc += args[0].length;
-	val2 = get_args(env->memory, pc, args[1].length);
+	val2 = (int16_t)get_args(env->memory, pc, args[1].length);
 	pc += args[1].length;
 	reg = get_args(env->memory, pc, args[2].length);
 	process->reg[reg - 1] = val1 & val2;
